Made updates to comments.
