+++ app/javascript/components/pages/Home.js
@@ -1,18 +1,9 @@
-import React, { Component} from "react";
-import {
-  Card,
-  Button,
-  CardTitle,
-  Row,
-  Col,
-  NavLink,
-} from "reactstrap";
+import React, { Component } from "react";
+import { Card, Button, CardTitle, Row, Col, NavLink } from "reactstrap";
 
 class Home extends Component {
   render() {
     let { drinks } = this.props;
-    //deconstruct ^
-    // console.log(this.props.state)
     return (
       <div className="main-body">
         <h2>Cocktails</h2>
@@ -24,7 +15,9 @@ class Home extends Component {
                   <Col sm="6">
                     <Card body>
                       <NavLink href={`/DrinkShow/${drink._id}`}>
-                        <CardTitle tag="h5">Cocktail {drink.drinkName}</CardTitle>
+                        <CardTitle tag="h5">
+                          Cocktail {drink.drinkName}
+                        </CardTitle>
                       </NavLink>
                       <NavLink href={`/DrinkShow/${drink._id}`}>
                         <Button>Show More</Button>
